Refactor to support multiple data sets
Concatenating time series and fitting the models is not correct,
because it constructs windows out of multiple time series. Instead,
one needs to construct design matrix and response vectors for each
time series and concatenate those. A preparedata method is added that
does this, and the fit methods for KNNModel and kMeansModel continue
to work without changes.

Tests are added to make sure that vectors of vectors work.